[Feat] 기능 테스트 확인을 위한 코드
